+++ .gitignore
@@ -12,6 +12,14 @@
 .spyderproject
 build
 
+rg/*
+!rg/__init__.py
+!rg/nifty_rg.py
+!rg/nifty_power_conversion_rg.py
+!rg/gfft_rg.py
+!rg/powerspectrum.py
+!rg/smoothing.py
+
 demos/*
 !demos/__init__.py
 !demos/demos_core.py
